Design patterns state pattern

+++ src/main/java/crackingcodeinterview/JumpNStairs.java
@@ -16,24 +16,25 @@ public class JumpNStairs {
         System.out.println(countStairs1(3));
     }
 
-    private static long countStairs1(long n){
-        long[] arr = new long[(int) (n + 1)];
+    private static int countStairs1(int n){
+        int[] arr = new int[n + 1];
         Arrays.fill(arr, -1);
         return countStairs1(n, arr);
     }
 
-    //For each recursive call(sub-problem), If we've seen this value of n before, return the cached value. Each time we compute a fresh value,
+    //For each recursive call(sub-problem), If we've seen this value of n before, return the cached value.
+    // Each time we compute a fresh value,
     //add it to the cache.
-    private static long countStairs1(long n, long[] arr) {
+    private static int countStairs1(int n, int[] arr) {
         if(n < 0){
             return 0;
         }else if( n == 0){
             return 1;
-        }else if(arr[(int)n] > -1){
-            return arr[(int)n];
+        }else if(arr[n] > -1){
+            return arr[n];
         }else {
-            arr[(int)n] = countStairs1(n - 1, arr) + countStairs1(n - 2, arr) + countStairs1(n - 3, arr);
-            return arr[(int) n];
+            arr[n] = countStairs1(n - 1, arr) + countStairs1(n - 2, arr) + countStairs1(n - 3, arr);
+            return arr[ n];
         }
     }
